Faliure model tests done

+++ models/skiEquipmentModelMysql.js
@@ -241,7 +241,8 @@ async function editUser(id, userType, username, password, firstName, lastName, c
     if (!validate.isValidInteger(id)) {
         throw new UserDataError("Invalid id");
     }
-    if (getUserById(id).length == 0) {
+    let result = await getUserById(id);
+    if (result.length == 0) {
         throw new UserDataError("Invalid id");
     }
     if (!validate.isValidUserType(userType)) {
@@ -250,7 +251,7 @@ async function editUser(id, userType, username, password, firstName, lastName, c
     if (!validate.isValidAlphanumeric(username)) {
         throw new UserDataError("Invalid username");
     }
-    else if (await checkIfUsernameIsTaken(username)) {
+    if (await checkIfUsernameIsTaken(username)) {
         throw new UserDataError("Username already taken");
     }
     if (!validate.isValidPassword(password)) {
@@ -283,6 +284,11 @@ async function editUser(id, userType, username, password, firstName, lastName, c
  * @throws SystemError if there is an error in the database while deleting
  */
 async function deleteUser(id) {
+    let result = await getUserById(id);
+    if (result == null) {
+        return result;
+    }
+
     if (!validate.isValidInteger(id)) {
         throw new UserDataError("Invalid id");
     }
@@ -364,13 +370,16 @@ async function addItem(name, description, itemCost, itemType, rentalState) {
     if (!validate.isValidAlphanumeric(name)) {
         throw new UserDataError("Invalid name");
     }
+    if (!validate.isValidDescription(description)) {
+        throw new UserDataError("Invalid name");
+    }
     if (!validate.isValidDecimal(itemCost)) {
         throw new UserDataError("Invalid cost");
     }
     if (!validate.isValidItemType(itemType)) {
         throw new UserDataError("Invalid item type");
     }
-    if (rentalState != 0 && rentalState != 1) {
+    if (!validate.isValidRentalState(rentalState)) {
         throw new UserDataError("Invalid rental state");
     }
 
@@ -443,6 +452,11 @@ async function editItemRentalState(id, rentalState) {
  * @param {*} id The id of the item to be deleted
  */
 async function deleteItem(id) {
+    let result = await getItemById(id);
+    if (result == null) {
+        return result;
+    }
+
     if (!validate.isValidInteger(id)) {
         throw new UserDataError("Invalid id");
     }
@@ -452,6 +466,7 @@ async function deleteItem(id) {
             logger.error(error)
             throw new SystemError("Error deleting item");
         });
+
 }
 
 /**
@@ -525,6 +540,12 @@ async function editItemType(id, name) {
         throw new UserDataError("Name already taken");
     }
 
+    let result = await getItemTypeById(id);
+
+    if (result == null) {
+        throw new UserDataError("Error ItemType not in database");
+    }
+
     const sqlQuery = 'UPDATE itemTypes SET name = \'' + name + '\' WHERE id = ' + id;
     await connection.execute(sqlQuery)
         .catch((error) => {
@@ -539,10 +560,16 @@ async function editItemType(id, name) {
  * @param {*} id The id of the item type to be deleted
  */
 async function deleteItemType(id) {
+    let result = await getItemTypeById(id);
+
+    if (result == null) {
+        throw new UserDataError("Error ItemType not in database");
+    }
+
     if (!validate.isValidInteger(id)) {
         throw new UserDataError("Invalid id");
     }
-    const sqlQuery = 'DELETE FROM itemTypes WHERE id = ' + id;
+    sqlQuery = 'DELETE FROM itemTypes WHERE id = ' + id;
     await connection.execute(sqlQuery)
         .catch((error) => {
             logger.error(error);
@@ -551,8 +578,24 @@ async function deleteItemType(id) {
 }
 
 /**
+ * Returns an item from database
+ * Used to check if an id is in the database
+ * @param {*} id The id of the item to be returned
+ */
+async function getItemTypeById(id) {
+    const sqlQuery = 'SELECT * FROM itemTypes WHERE id = \'' + id + '\'';
+    const result = await connection.execute(sqlQuery)
+        .catch((error) => {
+            logger.error(error)
+            throw new SystemError("Error getting item type");
+        });
+    return result[0][0] ? result[0][0] : null;
+}
+
+/**
  * Returns an item type from the database
  * Used to check if an item type name already exists
+ * @param {*} name The name of the item type to be returned
  */
 async function getItemTypeByName(name) {
     if (!validate.isValidAlphanumeric(name)) {
@@ -650,5 +693,6 @@ module.exports = {
     deleteItemType,
     getItemTypeByName,
     getAllItemTypes,
-    getConnection
+    getConnection,
+    getItemTypeById
 }
